# Reduce block fragmentation