Change `NaiveBlockAllocator` to use a priority queue so that we always allocate the lowest block id first. 

This further increases the performance of contiguous paged attention.

- [ ] Add an option or env variable to enable/disable this behavior. (Not sure if this is necessary)