@xuechendi 2024-10-25
PR424 vs PR424+PR426
|            | Gaudi3-1.18.0-524 #424 |          |                  | Gaudi3-1.18.0-524 #424 + #426 |          |                  |
| :--------- | ---------------------- | -------- | ---------------- | ----------------------------- | -------- | ---------------- |
| Batch size | TTFT (ms)              | TPOT(ms) | Thrput (tokes/s) | TTFT (ms)                     | TPOT(ms) | Thrput (tokes/s) |
| 128        | 1548                   | 35       | 3446.87          | 1530                          | 32       | 3795.67          |
| 96         | 1172                   | 33       | 2774             | 1162                          | 26       | 3422.33          |
| 64         | 801                    | 31       | 2003.72          | 789                           | 21       | 2805.18          |