@xuechendi 2024-10-30
@michalkuligowski , please help to review this PR. I see PR 433 is merged. I think PR426 is necessary for the performance.

Test on BS=128, input_len=1024, max_new_tokens=1024  with benchmark_throughput.py
```
VLLM_CONTIGUOUS_PA='true' VLLM_PA_SOFTMAX_IMPL='wsum' python benchmarks/benchmark_throughput.py \
        --model ${model} \
        --device hpu \
        --backend vllm \
        --num-prompts ${bs} \
        --input_len ${in_len} \
        --output_len ${out_len} \
        --max_model_len ${total_len} \
        --dtype bfloat16 \
        --num_scheduler_steps 1 \
        --gpu-memory-util 0.9
```

current habana_main - d3257b21
```
['First Token: bs == 2, avg_time is 55.641567187365126 msecs, counts is 64\n']
['Next Token: bs == 128, avg_time is 29.573468785173418 msecs, counts is 1024\n']
['Prepare Input: bs == torch.Size([128]), avg_time is 4.008329316459225 msecs, counts is 1024\n']
====== Warmup ======
Throughput: 3.31 requests/s,
duration: 38.64 secs
Total Throughput: 6783.83 tokens/s
Input Throughput: 3391.91 tokens/s
Output Throughput: 3391.91 tokens/s
```

With PR426
```
['First Token: bs == 2, avg_time is 55.512639796859276 msecs, counts is 64\n']
['Next Token: bs == 128, avg_time is 26.642957841762183 msecs, counts is 1024\n']
['Prepare Input: bs == torch.Size([128]), avg_time is 3.858626422857725 msecs, counts is 1024\n']

====== Result======
Throughput: 3.58 requests/s,
duration: 35.74 secs
Total Throughput: 7333.89 tokens/s
Input Throughput: 3666.94 tokens/s
Output Throughput: 3666.94 tokens/s
```