@xuechendi 2024-10-30
A side-by-side analysis as below:
Conclusion, PR426 effectively reduces fragment after returning blocks and re-allocate.

After warmup, we can see left side with PR426, decode block_mapping start from shape 644 and gradually increase to 1540
While, with current habana_main, decode block_mapping start from shape 1540(Due to lots of fragment), and which leads to an unnecessary larger input tensor to be processed by flat_pa => and results in higher next_token_latency.
  
Left side: with PR426                 <=========================>        Right Side: current habana_main
![image](https://github.com/user-attachments/assets/5e430953-fc3f-4591-8399-cd20b856983e)